# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "CommandDetailDto": {
        "properties": {
          "command": {
            "$ref": "#/components/schemas/CommandDto"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/CommandResultDto"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "command",
          "results"
        ],
        "title": "CommandDetailDto",
        "type": "object"
      },
      "CommandDto": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "issued_by": {
            "title": "Issued By",
            "type": "string"
          },
          "params": {
            "title": "Params",
            "type": "object"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "idempotency_key",
          "type",
          "status",
          "params",
          "issued_by",
          "created_at",
          "updated_at",
          "expires_at",
          "reason",
          "attempts"
        ],
        "title": "CommandDto",
        "type": "object"
      },
...
      "ExecutionDto": {
        "properties": {
          "broker_trade_id": {
            "title": "Broker Trade Id",
            "type": "string"
          },
          "fees": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fees"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "string"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "price": {
            "title": "Price",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          }
        },
        "required": [
          "id",
          "broker_trade_id",
          "order_id",
          "instrument_id",
          "side",
          "quantity",
          "price",
          "fees",
          "ts"
        ],
        "title": "ExecutionDto",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "passcode": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Passcode",
            "type": "string"
          }
        },
        "required": [
          "passcode"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "OrderDetailDto": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/OrderEventDto"
            },
            "title": "Events",
            "type": "array"
          },
          "order": {
            "$ref": "#/components/schemas/OrderDto"
          }
        },
        "required": [
          "order",
          "events"
        ],
        "title": "OrderDetailDto",
        "type": "object"
      },
      "OrderDto": {
        "properties": {
          "average_price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Price"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "filled_quantity": {
            "title": "Filled Quantity",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "string"
          },
          "limit_price": {
            "title": "Limit Price",
            "type": "string"
          },
          "order_type": {
            "title": "Order Type",
            "type": "string"
          },
          "ordertag": {
            "title": "Ordertag",
            "type": "string"
          },
          "parent_order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Order Id"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "reprice_count": {
            "title": "Reprice Count",
            "type": "integer"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "signal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Id"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "status_message": {
            "title": "Status Message",
            "type": "string"
          },
          "strategy_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Run Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ordertag",
          "instrument_id",
          "side",
          "order_type",
          "quantity",
          "filled_quantity",
          "limit_price",
          "average_price",
          "state",
          "strategy_run_id",
          "signal_id",
          "parent_order_id",
          "reprice_count",
          "status_message",
          "created_at",
          "updated_at"
        ],
        "title": "OrderDto",
        "type": "object"
      },
...
      "PositionDto": {
        "properties": {
          "average_price": {
            "title": "Average Price",
            "type": "string"
          },
          "fees": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fees"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "string"
          },
          "net_quantity": {
            "title": "Net Quantity",
            "type": "integer"
          },
          "realised_pnl": {
            "title": "Realised Pnl",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "instrument_id",
          "net_quantity",
          "average_price",
          "realised_pnl",
          "fees",
          "updated_at"
        ],
        "title": "PositionDto",
        "type": "object"
      },
...
      "RiskDto": {
        "properties": {
          "limits": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Limits",
            "type": "object"
          },
          "recent_rejections": {
            "items": {
              "$ref": "#/components/schemas/RiskEventDto"
            },
            "title": "Recent Rejections",
            "type": "array"
          }
        },
        "required": [
          "limits",
          "recent_rejections"
        ],
        "title": "RiskDto",
        "type": "object"
      },
...
      "SubmitCommandRequest": {
        "description": "`idempotency_key` is client-generated: the same key can never execute twice.",
        "properties": {
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "params": {
            "title": "Params",
            "type": "object"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "idempotency_key",
          "type"
        ],
        "title": "SubmitCommandRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "expires_at"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Emporos control plane",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login"
      }
    },
    "/commands": {
      "get": {
        "operationId": "commands_commands_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CommandDto"
                  },
                  "title": "Response Commands Commands Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commands"
      },
      "post": {
        "description": "Record the command and return at once: 202, status PENDING. The outcome is read from\n`GET /commands/{id}` or the stream \u2014 never assumed.",
        "operationId": "submit_commands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit"
      }
    },
    "/commands/{command_id}": {
      "get": {
        "operationId": "command_commands__command_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "command_id",
            "required": true,
            "schema": {
              "title": "Command Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandDetailDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Command"
      }
    },
    "/executions": {
      "get": {
        "operationId": "executions_executions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExecutionDto"
                  },
                  "title": "Response Executions Executions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Executions"
      }
    },
...
    "/orders": {
      "get": {
        "operationId": "orders_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "instrument_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Instrument Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  },
                  "title": "Response Orders Orders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Orders"
      }
    },
    "/orders/{order_id}": {
      "get": {
        "operationId": "order_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Order"
      }
    },
...
    "/positions": {
      "get": {
        "operationId": "positions_positions_get",
        "parameters": [
          {
            "in": "query",
            "name": "open_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Open Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PositionDto"
                  },
                  "title": "Response Positions Positions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Positions"
      }
    },
    "/risk": {
      "get": {
        "operationId": "risk_risk_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Risk"
      }
    },
...
    "/strategies/{name}/graduation": {
      "get": {
        "operationId": "strategy_graduation_strategies__name__graduation_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraduationDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Strategy Graduation"
      }
    },
...
    "/system/events": {
      "get": {
        "operationId": "system_events_system_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SystemEventDto"
                  },
                  "title": "Response System Events System Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "System Events"
      }
    },
    "/system/reconciliations": {
      "get": {
        "operationId": "reconciliations_system_reconciliations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReconciliationDto"
                  },
                  "title": "Response Reconciliations System Reconciliations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconciliations"
      }
    }
  }
}